move kappa calculation out from preprocessing step
